return to using CentroidX
Remove the usage of the radius in the effective area calculation. We should never need to use the radius and it is good to make the code less robust to issues like the one we saw here with the camera rotation. This way we will notice issues like that in the future.