Change push-to-package repository conditions
Packages where pushed to the repository each time a pull request is
issued or updated, but not after a merge to main.

Changed this so that only successfully merged pull request result in a
package push.

This does not change when this workflows are triggered; only the
'package push' conditions are changed.